AFKMud 2.0.3
Recommit version 2.0.3 for change history.

+++ src/act_info.cpp
@@ -5,7 +5,7 @@
  *                /-----\  |      | \  |  v  | |     | |  /                 *
  *               /       \ |      |  \ |     | +-----+ +-/                  *
  ****************************************************************************
- * AFKMud Copyright 1997-2007 by Roger Libiez (Samson),                     *
+ * AFKMud Copyright 1997-2008 by Roger Libiez (Samson),                     *
  * Levi Beckerson (Whir), Michael Ward (Tarl), Erik Wolfe (Dwip),           *
  * Cameron Carroll (Cam), Cyberfox, Karangi, Rathian, Raine,                *
  * Xorith, and Adjani.                                                      *
@@ -323,7 +323,7 @@ void show_condition( char_data * ch, char_data * victim )
    int percent;
 
    if( victim->max_hit > 0 )
-      percent = ( 100 * victim->hit ) / victim->max_hit;
+      percent = ( int )( ( 100.0 * ( double )( victim->hit ) ) / ( double )( victim->max_hit ) );
    else
       percent = -1;
